Move auth components into separate folder
Also, document AuthenticationComponent.
